requirements.txt
Pillow==5.0.0
django==2.0.2
django-ckeditor==5.4.0
pandas
folium
social-auth-app-django
ipython